Extracted BaseServiceModel and Dao for finer grained control over write operations.
